2000 versions of Unity later...
So: The code that I was working on doesn't quite work properly. That being said, it's now on a newer version of unity, and for whatever reason my computer has defaulted to opening CS files with wordpad. So as I attempt to find a way to get it to use monodevelop as I have been, I'll let you all have access to this. Additionally, if you would like to use my previous code, you are more than welcome to.